fix: clear up template build errors
