+++ packages/vue-template/package.json
@@ -4,17 +4,15 @@
   "private": true,
   "type": "module",
   "scripts": {
-    "dev": "vite",
     "build": "run-p type-check \"build-only {@}\" --",
+    "build-only": "vite build",
+    "dev": "vite",
+    "lint": "eslint",
+    "lint:fix": "eslint --fix",
     "preinstall": "node preinstall.cjs",
     "prepare": "node prepare.cjs",
     "preview": "vite preview",
-    "test:unit": "vitest",
-    "test:e2e": "playwright test",
-    "build-only": "vite build",
-    "type-check": "vue-tsc --build",
-    "lint": "eslint",
-    "lint:fix": "eslint --fix"
+    "type-check": "vue-tsc --build"
   },
   "dependencies": {
     "@tailwindcss/vite": "^4.1.10",
@@ -30,8 +28,10 @@
     "@headlessui/tailwindcss": "^0.2.2",
     "@headlessui/vue": "^1.7.23",
     "@ignt/vue-library": "^1.0.5",
+    "@keplr-wallet/types": "^0.12.247",
     "@stylistic/eslint-plugin": "^5.0.0",
     "@tsconfig/node22": "^22.0.2",
+    "@types/crypto-js": "^4.2.2",
     "@types/jsdom": "^21.1.7",
     "@types/node": "^24.0.4",
     "@vitejs/plugin-vue": "^6.0.0",
