+++ packages/vue-template/src/def-composables/useTransactions.ts
@@ -13,9 +13,9 @@ export const useTransactions = () => {
   const RECEIVED_EVENT = computed<string>(
     () => `transfer.recipient='${address.value}'`
   );
-  const sentQuery = ServiceGetTxsEvent({ query: SENT_EVENT.value }, {}, 100);
+  const sentQuery = ServiceGetTxsEvent({ query: SENT_EVENT.value } as Parameters<typeof ServiceGetTxsEvent>[0], {}, 100);
   const receivedQuery = ServiceGetTxsEvent(
-    { query: RECEIVED_EVENT.value },
+    { query: RECEIVED_EVENT.value } as Parameters<typeof ServiceGetTxsEvent>[0],
     {},
     100
   );
