+++ packages/vue-template/tsconfig.app.json
@@ -4,7 +4,7 @@
   "exclude": ["src/**/__tests__/*"],
   "compilerOptions": {
     "tsBuildInfoFile": "./node_modules/.tmp/tsconfig.app.tsbuildinfo",
-
+    "typeRoots": ["./node_modules/@types", "./"],
     "paths": {
       "@/*": ["./src/*"]
     }
